Mudanças na geração no output
- Output gerado agora por método e não por quantidade.
- TODO: Revisão no makefile de testes. é preciso rodar duas vezes para não dar erro.

+++ test/unit_test.c
@@ -12,7 +12,7 @@ UnitTest createUnitTest(void *vector, int n, int size, Comparator comparator, Ev
 }
 
 
-void evaluateUnitTest(UnitTest * unitTest, double time) {
+void evaluateUnitTest(UnitTest * unitTest, long double time) {
     unitTest->result = unitTest->evaluator( unitTest->vector, 
                                             unitTest->n, 
                                             unitTest->size,
